Move missions list to separate component.
